web/pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["Manuel Antonio"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
Markdown = "^3.4.1"
django-filter = "^22.1"
drf-yasg = "^1.21.4"
dj-rest-auth = "^2.2.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
